Working full impulse solver

+++ PhysiXSharp.Core/Physics/Data/CollisionResolution.cs
@@ -9,10 +9,10 @@ public sealed class CollisionResolution
     public readonly Rigidbody RigidbodyB;
     public readonly Vector VelocityChangeA;
     public readonly Vector VelocityChangeB;
-    public readonly float AngularVelocityChangeA;
-    public readonly float AngularVelocityChangeB;
+    public readonly double AngularVelocityChangeA;
+    public readonly double AngularVelocityChangeB;
 
-    public CollisionResolution(Rigidbody rigidbodyA, Rigidbody rigidbodyB, Vector velocityA, Vector velocityB, float angularA, float angularB)
+    public CollisionResolution(Rigidbody rigidbodyA, Rigidbody rigidbodyB, Vector velocityA, Vector velocityB, double angularA, double angularB)
     {
         RigidbodyA = rigidbodyA;
         RigidbodyB = rigidbodyB;
